+++ software/system/roneLib/src/Behaviors/behaviorSystem.h
@@ -50,7 +50,7 @@ void motorSetBeh(Beh* behPtr);
  * @param behPtr for behavior
  * @returns translational velocity of input behavior
  */
-uint8 behGetTv(Beh* behPtr);
+int32 behGetTv(Beh* behPtr);
 
 
 /**
@@ -90,7 +90,7 @@ void behSetRv(Beh* behPtr, int32 rv);
  *  @param behPtr for behavior
  *  @returns the rotational velocity of the input behavior
  */
-uint8 behGetRv(Beh* behPtr);
+int32 behGetRv(Beh* behPtr);
 
 
 /**
